Refactor #11: Moved all IOManger code to InputManager since IOManager was not handling outputs (keeping IOManager for merge with main)
