+++ src/main/java/ca/mcmaster/se2aa4/mazerunner/Main.java
@@ -1,13 +1,5 @@
 package ca.mcmaster.se2aa4.mazerunner;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-
-import org.apache.commons.cli.Options;
-import org.apache.commons.cli.CommandLine;
-import org.apache.commons.cli.CommandLineParser;
-import org.apache.commons.cli.DefaultParser;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -17,14 +9,14 @@ public class Main {
     private static final Logger logger = LogManager.getLogger();
 
     public static void main(String[] args) {
-        IOManager ioManager = new IOManager(args);
+        InputManager inputManager = new InputManager(args);
         // System.out.println("** Starting Maze Runner");
         logger.info("** Starting Maze Runner");
         try {
             // System.out.println(ioManager.getFilePath());
             // System.out.println(ioManager.getUserPath());
 
-            RectangleMaze maze = new RectangleMaze(ioManager.getFilePath());
+            RectangleMaze maze = new RectangleMaze(inputManager.getFilePath());
             // maze.testDisplay();
             // logger.info("Entry: {}", maze.getEntryPos());
             // logger.info("Exit: {}", maze.getExitPos());
@@ -32,7 +24,7 @@ public class Main {
             // logger.info("**** Computing path");
             Path path = new Path(maze);
             System.out.println("PATH: " + path.getPath());
-            
+
             
         } catch(Exception e) {
             System.err.println(e);
